+++ projects/crypto-currency/destemidos/App/src/pages/Markets/style.css
@@ -1,0 +1,20 @@
+
+.container{
+  max-width: 900px;
+  margin: 0 auto;
+  align-self: center;
+  align-items: center;
+  height: 100vh;
+}
+
+.SliderView{
+  width: 900px;
+  text-align: right;
+  margin: 0 auto;
+}
+
+.SliderTitle{
+  max-width: 900px;
+  text-align: left;
+  margin: 0 auto;
+}
